split giant test_views file into 4 files
